pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "cityzero"
version = "0.2.2"
description = "CLI tool to bulk-download street-level imagery from Mapillary"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Stiven" }]
dependencies = [
    "folium>=0.17",
    "mapillary>=1.0",
    "piexif>=1.1",
    "python-dotenv>=1.0",
    "questionary>=2.0",
    "requests>=2.31",
    "tqdm>=4.66",
]

[project.urls]
Homepage = "https://github.com/SomeoneElseSt/mapillary-dl"
Contact = "https://stiven.me"

[project.scripts]
cityzero = "cli:main"

[tool.setuptools]
py-modules = ["cli", "config", "database", "downloader"]
